uboot: add single command nand flash burn function
Summary:
for the convinience of burning nand flash, added a command for
burning whole image onto nand flash. The syntax for using the command is:
burn_nand [addr] [size]
in which, addr is the address of image in DDR, size is the size of the
decompressed image

Test Plan:
In both emmc and nand boot uboot, write to nand flash successfully.
In emmc boot, should be able to mount both sys and rootfs partition.
In nand boot, should be reset into Kernel successfully

Reviewers: huiliang.xu, xiaofeng.ling

Reviewed By: huiliang.xu, xiaofeng.ling

Differential Revision: https://cr.hobot.cc/D20893