env: mmc: correct the offset returned by mmc_offset_try_partition
Summary:
The output of the function mmc_offset_try_partition must be a
byte offset in mmc and not a multiple of blksz.

This function is used in mmc_offset(), called by mmc_get_env_addr()
and the offset is used in write_env(), erase_env() and read_env().

In these function, blk_start = offset / mmc->read_bl_len
or /write_bl_len so this offset is not a multiple of blksz.

Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>

Test Plan: NA

Reviewers: yunqian.wang, xiaofeng.ling, dinggao.pan

Reviewed By: yunqian.wang, xiaofeng.ling, dinggao.pan

Subscribers: ming.yu

Differential Revision: https://cr.hobot.cc/D115611